1. Synchronize with ROC repository 2. Prepare for integration into ROClib 3. Better handling of ROOT in the DABC
git-svn-id: https://subversion.gsi.de/dabc/trunk@662 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ plugins/roc/src/FEBboard.cxx
@@ -60,7 +60,7 @@ void roc::FEBboard::reset_I2C()
 //-------------------------------------------------------------------------------
 nxyter::Chip &roc::FEBboard::NX(int index)
 {
-   if (index>=(int)fNX.size()) throw std::out_of_range("FATAL ERROR: You tried to access an nXYTER Number, that is not available!\r\n");
+   if (index>=fNX.size()) throw std::out_of_range("FATAL ERROR: You tried to access an nXYTER Number, that is not available!\r\n");
    return *fNX[index];
 }
 //-------------------------------------------------------------------------------
